aws/sam/mads-swap-app/swap_simulation/requirements.txt
requests
aws-psycopg2
pandas